feat: Refactor time base to separate raw and corrected microseconds, add IMU temperature data, and improve IMU rate logging.

+++ stm32/Drivers/USER/cc/m9n.cpp
@@ -158,7 +158,7 @@ void M9N::FlashConfig(uint32_t current_baud) {
 
   // Small delay for UART to stabilize (100ms)
   auto &time = System::GetInstance().Time();
-  uint64_t start = time.Micros();
+  uint32_t start = time.Micros();
   while ((time.Micros() - start) < MILLIS_TO_MICROS(100)) {
     // Busy wait
   }
